k8s: merged cache.go into client.go
